+++ cli/cluster/cmd/replicaset.go
@@ -10,6 +10,7 @@ import (
 	"github.com/tarantool/go-tarantool/v2"
 	"github.com/tarantool/tt/cli/replicaset"
 	libcluster "github.com/tarantool/tt/lib/cluster"
+	"github.com/tarantool/tt/lib/connect"
 	clientv3 "go.etcd.io/etcd/client/v3"
 )
 
@@ -107,7 +108,7 @@ func pickPatchKey(keys []string, force bool, pathMsg string) (int, error) {
 func createDataCollectorAndKeyPublisher(
 	collectors libcluster.DataCollectorFactory,
 	publishers libcluster.DataPublisherFactory,
-	opts UriOpts, connOpts connectOpts) (
+	opts connect.UriOpts, connOpts connectOpts) (
 	libcluster.DataCollector, replicaset.DataPublisher, func(), error) {
 	prefix, key, timeout := opts.Prefix, opts.Key, opts.Timeout
 	var (
@@ -154,7 +155,7 @@ func createDataCollectorAndKeyPublisher(
 
 // Promote promotes an instance by patching the cluster config.
 func Promote(uri *url.URL, ctx PromoteCtx) error {
-	opts, err := ParseUriOpts(uri)
+	opts, err := connect.ParseUriOpts(uri)
 	if err != nil {
 		return fmt.Errorf("invalid URL %q: %w", uri, err)
 	}
@@ -201,7 +202,7 @@ type DemoteCtx struct {
 
 // Demote demotes an instance by patching the cluster config.
 func Demote(uri *url.URL, ctx DemoteCtx) error {
-	opts, err := ParseUriOpts(uri)
+	opts, err := connect.ParseUriOpts(uri)
 	if err != nil {
 		return fmt.Errorf("invalid URL %q: %w", uri, err)
 	}
@@ -248,7 +249,7 @@ type ExpelCtx struct {
 
 // Expel expels an instance by patching the cluster config.
 func Expel(uri *url.URL, ctx ExpelCtx) error {
-	opts, err := ParseUriOpts(uri)
+	opts, err := connect.ParseUriOpts(uri)
 	if err != nil {
 		return fmt.Errorf("invalid URL %q: %w", uri, err)
 	}
@@ -302,7 +303,7 @@ type RolesChangeCtx struct {
 
 // ChangeRole adds/removes a role by patching the cluster config.
 func ChangeRole(uri *url.URL, ctx RolesChangeCtx, action replicaset.RolesChangerAction) error {
-	opts, err := ParseUriOpts(uri)
+	opts, err := connect.ParseUriOpts(uri)
 	if err != nil {
 		return fmt.Errorf("invalid URL %q: %w", uri, err)
 	}